+++ portable/src/implementation/portable/algorithm_safe.rs
