+++ portable/src/implementation/portable/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod algorithm_portable;
+pub(crate) mod algorithm_safe;
 #[cfg(any(not(feature = "simd256"), feature = "public_imp"))]
 pub(crate) mod simd128;
 #[cfg(any(feature = "simd256", feature = "public_imp"))]
